update PROJECT-USER app setting handling (#1128, #1129)

+++ projectroles/management/commands/cleanappsettings.py
@@ -16,6 +16,7 @@ DEFINITION_NOT_FOUND_MSG = 'definition not found'
 ALLOWED_TYPES_MSG = 'does not match allowed types'
 DELETE_PREFIX_MSG = 'Deleting "{}" from project "{}": '
 DELETE_PROJECT_TYPE_MSG = 'project type "{}" {}: {}'
+DELETE_SCOPE_MSG = 'user {} has no role in project'
 
 
 def get_setting_str(db_setting):
@@ -68,6 +69,19 @@ class Command(BaseCommand):
                         ALLOWED_TYPES_MSG,
                         definition.get('project_types', ['PROJECT']),
                     )
+                )
+                s.delete()
+
+        db_settings = AppSetting.objects.filter(
+            project__isnull=False, user__isnull=False
+        )
+        for s in db_settings:
+            if not s.project.get_role(s.user):
+                logger.info(
+                    DELETE_PREFIX_MSG.format(
+                        get_setting_str(s), s.project.title
+                    )
+                    + DELETE_SCOPE_MSG.format(s.user.username)
                 )
                 s.delete()
         logger.info(END_MSG)
